pyproject.toml
[tool.poetry]
name = "course paper"
version = "0.1.0"
description = ""
authors = ["Ilya Semak <ladysemak2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
dt = "^1.1.64"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
python-dotenv = "^1.0.1"
pandas-stubs = "^2.2.2.240807"
types-requests = "^2.32.0.20240712"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.11.1"
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
